+++ cryptoconseils-frontend/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import Home from './Home'
+import Home from './Home/Home'
 class App extends Component {
   render() {
     return (
